+++ tests/conftest.py
@@ -44,7 +44,7 @@ from inmanta.server.bootloader import InmantaBootloader
 from inmanta.server import SLICE_AGENT_MANAGER
 from inmanta.export import cfg_env, unknown_parameters
 import traceback
-from tornado import process
+from tornado import process, netutil
 import asyncio
 from tornado.platform.asyncio import AnyThreadEventLoopPolicy
 import asyncpg
@@ -53,6 +53,8 @@ import sys
 import pkg_resources
 from typing import Optional, Dict
 from inmanta import protocol
+import pyformance
+from pyformance.registry import MetricsRegistry
 
 
 asyncio.set_event_loop_policy(AnyThreadEventLoopPolicy())
@@ -91,11 +93,16 @@ def deactive_venv():
     pkg_resources.working_set = pkg_resources.WorkingSet._build_master()
 
 
+def reset_metrics():
+    pyformance.set_global_registry(MetricsRegistry())
+
+
 def reset_all_objects():
     resources.resource.reset()
     export.Exporter.reset()
     process.Subprocess.uninitialize()
     asyncio.set_child_watcher(None)
+    reset_metrics()
     # No dynamic loading of commands at the moment, so no need to reset/reload
     # command.Commander.reset()
     handler.Commander.reset()
@@ -135,6 +142,13 @@ def no_agent_backoff():
     inmanta.agent.agent.GET_RESOURCE_BACKOFF = 0
     yield
     inmanta.agent.agent.GET_RESOURCE_BACKOFF = backoff
+
+
+@pytest.fixture()
+def free_socket():
+    sock = netutil.bind_sockets(0, "127.0.0.1", family=socket.AF_INET)[0]
+    yield sock
+    sock.close()
 
 
 def get_free_tcp_port():
@@ -210,6 +224,8 @@ async def agent(server, environment):
 async def server(inmanta_config, postgres_db, database_name):
     # fix for fact that pytest_tornado never set IOLoop._instance, the IOLoop of the main thread
     # causes handler failure
+
+    reset_metrics()
 
     state_dir = tempfile.mkdtemp()